75 - Добавить OAuth-авторизацию в проект (исправлена логика обработки code)
